refactor(stella-tui): move the ? overlay out of deck_render.rs
deck_render.rs is a grandfathered god file at 1514 lines against a 1518
ceiling — four lines of headroom. The metric rows SPEC 5 sends behind ? are
forty, so the move is #4188's enabling step rather than incidental to it.

The cut follows the concern: help_row, tab_shortcuts, GLOBAL_SHORTCUTS and
render_help all answer 'what does ? draw', and nothing else in deck_render
reads any of them. Pure move, no behaviour change; render_help gains
pub(super) so its one call site still reaches it.

deck_render.rs 1514 -> 1315. No baseline entry added, no ceiling raised.

Refs #4188
