+++ crates/stella-tui/src/deck_render.rs
@@ -1308,205 +1308,13 @@ fn render_composer_footer(
     render_right(right, area, buf);
 }
 
-/// One aligned `key → description` row of the help overlay. The key column is
-/// padded to a fixed width so the descriptions line up into a scannable
-/// second column.
-fn help_row(key: &str, desc: &str) -> Line<'static> {
-    Line::from(vec![
-        Span::styled(format!("  {key:<13} "), theme::accent()),
-        Span::styled(desc.to_string(), theme::body()),
-    ])
-}
 
-/// The shortcuts specific to one deck tab, as `(key, description)` pairs.
-/// Keyed off [`DeckTab`] so the overlay only ever shows keys that work where
-/// the user actually is — the per-tab handlers in `deck_ui` are the behavior
-/// these rows must mirror.
-fn tab_shortcuts(tab: DeckTab) -> &'static [(&'static str, &'static str)] {
-    match tab {
-        DeckTab::Session => &[
-            ("↑ ↓", "select a message · esc clears the selection"),
-            ("⇞ ⇟", "scroll the transcript"),
-            ("⌘[ / ⌘]", "jump to transcript start / end (⌃ works too)"),
-            ("ctrl-o", "expand/collapse the selected message (none: all)"),
-            ("ctrl-r", "expand/collapse all thinking"),
-            (
-                "ctrl-f",
-                "find in the transcript — ⏎ next · ctrl-p previous",
-            ),
-            ("ctrl-n / ctrl-p", "jump to the next / previous failure"),
-            ("ctrl-z", "fold the selected turn to one line (none: all)"),
-            ("↑", "with prompts queued: open the queue editor"),
-            ("←", "SESSIONS overlay — every session on this machine"),
-            ("→", "CONTEXT overlay — active skills + MCP servers"),
-            ("ctrl-g", "INSPECT — the context sent on any recorded call"),
-            (
-                "a / t / x",
-                "plan review: approve · trim · abort — one keypress decides",
-            ),
-            (
-                "r",
-                "plan review: refine — type what to change, ⏎ re-plans from it",
-            ),
-            (
-                "esc",
-                "plan review: abort (from the refine input: back out)",
-            ),
-        ],
-        DeckTab::Agents => &[
-            ("← →", "switch panes — executions / installed"),
-            ("↑ ↓", "select an agent"),
-            ("⏎", "open the selected agent's session"),
-            ("s / p / r", "stop · pause · restart the selected agent"),
-            ("⏎", "installed pane: edit · v versions · r reload"),
-            ("n", "new agent — drafted by the LLM"),
-        ],
-        DeckTab::Traces => &[
-            ("↑ ↓ ⇞ ⇟", "scroll the event log"),
-            ("f", "cycle the per-agent filter"),
-        ],
-        DeckTab::Graph => &[
-            ("← → ↑ ↓", "walk the neighborhood"),
-            ("/ or ⏎", "file picker — re-root on any indexed file"),
-        ],
-        DeckTab::Files => &[("↑ ↓", "select a file"), ("⏎", "open / close the diff")],
-        DeckTab::Skills => &[
-            ("← →", "switch panes"),
-            ("↑ ↓", "select a skill"),
-            ("space", "enable / disable"),
-            ("e", "edit the selected skill"),
-            ("p", "pin / unpin"),
-            ("n", "new skill — drafted by the LLM"),
-            ("ctrl-o", "preview"),
-            ("ctrl-x ×2", "delete (press twice to confirm)"),
-            ("type", "search skills"),
-        ],
-        DeckTab::Mcp => &[
-            ("↑ ↓", "select a server"),
-            ("space / e", "enable / disable"),
-            ("a", "authenticate (env credentials)"),
-            ("o", "OAuth login (http servers)"),
-            ("s", "search the registry"),
-            ("x", "remove the server"),
-            ("r", "refresh"),
-        ],
-        DeckTab::Issues => &[
-            ("↑ ↓", "select an issue"),
-            ("r", "refresh the list"),
-            ("/", "search the tracker"),
-            ("n", "new issue — tab cycles fields · ctrl-s creates"),
-            ("c", "comment on the selected issue"),
-            ("s", "set the selected issue's status"),
-            ("w", "start work on the selected issue"),
-        ],
-        DeckTab::Settings => &[
-            ("← →", "switch panes — agents / tools"),
-            ("e", "edit the pane you are on"),
-            ("t", "jump to the tool switches"),
-            (
-                "tab",
-                "in the editor: switch agent — global / default / worker / …",
-            ),
-            ("⏎", "in the editor: edit the selected row / pick a model"),
-            ("space", "in the editor: toggle the selected row"),
-            ("x", "in the editor: clear the selected row"),
-            ("s / S", "in the editor: save to user / project settings"),
-            ("r", "in the editor: reload from disk"),
-            ("esc", "in the editor: hand the keyboard back to the tab"),
-        ],
-    }
-}
-
-/// Deck-wide shortcuts that work on every tab.
-const GLOBAL_SHORTCUTS: &[(&str, &str)] = &[
-    ("tab / ⇧tab", "switch tabs"),
-    // Enter semantics are the inverse of the old chord-to-submit mapping (see
-    // `composer::classify_enter`): a bare ⏎ dispatches, a *modified* ⏎ breaks
-    // the line. The composer footer advertises the same pair — these rows must
-    // not drift from it.
-    // The parenthetical is load-bearing: the unqualified promise ("runs as its
-    // own agent") is what a reviewer read while a scope card was waiting, and
-    // the sidecar it describes is exactly what swallowed their answer.
-    (
-        "⏎",
-        "queue the prompt — mid-turn it runs as its own agent (unless a card waits)",
-    ),
-    ("⌘⏎ / ⌃⏎ / ⌥⏎", "insert a line break in the prompt"),
-    ("!cmd", "run a shell command NOW (skips the queue)"),
-    ("/", "slash commands — ↑↓ pick · tab completes · ⏎ runs"),
-    ("ctrl-v", "paste — a copied image is attached to the prompt"),
-    ("ctrl-t", "open the queue editor"),
-    ("ctrl-s", "PLAN — every step of the approved plan, in full"),
-    (
-        ">text",
-        "steer the running turn — lands at the next step boundary",
-    ),
-    ("esc", "steer: your draft + queue go INTO the running turn"),
-    (
-        "esc esc",
-        "cancel NOW & hold — nothing runs until your next prompt",
-    ),
-    ("ctrl-c", "quit stella"),
-];
-
-/// The help overlay: the active tab's keys first, then the deck-wide keys —
-/// one shortcut per line, key column aligned. Context-aware on purpose: only
-/// shortcuts that work on the tab the user is looking at are shown, so the
-/// overlay stays short enough to read at a glance. Opened by `?` (empty
-/// composer) or `/help`; scrolls with ↑/↓/⇞/⇟/Home/End on a short terminal;
-/// closes with esc/`q`/`?`.
-fn render_help(ui: &mut DeckUi, area: Rect, buf: &mut Buffer) {
-    let mut lines: Vec<Line<'static>> = Vec::new();
-    lines.push(Line::default());
-    lines.push(Line::from(Span::styled(
-        format!("  {} tab", ui.tab.title()),
-        theme::heading(),
-    )));
-    for (key, desc) in tab_shortcuts(ui.tab) {
-        lines.push(help_row(key, desc));
-    }
-    lines.push(Line::default());
-    lines.push(Line::from(Span::styled("  everywhere", theme::heading())));
-    for (key, desc) in GLOBAL_SHORTCUTS {
-        lines.push(help_row(key, desc));
-    }
-    lines.push(Line::default());
-    lines.push(Line::from(Span::styled(
-        "  letter & arrow hotkeys apply while the prompt box is empty",
-        theme::muted(),
-    )));
-
-    // Size the panel to its content, capped to the frame.
-    let w = area.width.min(68);
-    let h = area.height.min(lines.len() as u16 + 2);
-    let popup = Rect {
-        x: area.x + (area.width.saturating_sub(w)) / 2,
-        y: area.y + (area.height.saturating_sub(h)) / 2,
-        width: w,
-        height: h,
-    };
-    Clear.render(popup, buf);
-    let block = Block::default()
-        .borders(Borders::ALL)
-        .border_style(theme::accent())
-        .title(format!(" help — {} · esc close ", ui.tab.title()));
-    let inner = block.inner(popup);
-    block.render(popup, buf);
-    if inner.height == 0 || inner.width == 0 {
-        return;
-    }
-
-    let total = lines.len();
-    let height = inner.height as usize;
-    // Record viewport metrics for the pure key handler (`handle_help_key`) —
-    // when the panel is clipped, ↑/↓/⇞/⇟/Home/End scroll it.
-    ui.metrics.help_total = total;
-    ui.metrics.help_height = height;
-    let window = ui.help_scroll.window(total, height);
-    Paragraph::new(lines)
-        .scroll((window.start as u16, 0))
-        .render(inner, buf);
-}
+// The `?` overlay — SPEC 11's key sheet and SPEC 5's metric detail. Split out
+// rather than grown here: this file is a grandfathered god file closed to
+// growth, and the metric rows #4188 asks for do not fit in its four lines of
+// headroom. `help.rs`'s module doc carries the argument.
+mod help;
+use help::render_help;
 
 mod parked;
